requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
diffusers==0.31.0
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.10.0
future==1.0.0
h11==0.14.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
outcome==1.3.0.post0
packaging==24.2
pillow==11.0.0
PySocks==1.7.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
